requirements.txt
asn1crypto==1.5.1
attrs==22.2.0
autopep8==2.0.1
boto3==1.26.70
botocore==1.29.70
flake8==6.0.0
iniconfig==2.0.0
jmespath==1.0.1
mccabe==0.7.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pg8000==1.29.4
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv ==1.0.0
pytz==2022.7.1
s3transfer==0.6.0
scramp==1.4.4
six==1.16.0
urllib3==1.26.14
